+++ factors.h
@@ -2,7 +2,7 @@
 #define FACTORS_H
 #include <stdio.h>
 
-int square_root(int n);
-void factorize(int num);
+int square_root(long long int n);
+void factorize(long long int num);
 
 #endif /* FACTORS_H */
